extract shareable model logic into libraries
